refactor(config): move active_archive_root out of api.archive
Runtime-root resolution is not facade-specific -- it belongs to the
config ring, not polylogue.api. Adds active_archive_root(config) to
polylogue/config.py (the config ring's core module) and removes the
private _active_archive_root from polylogue/api/archive.py, keeping
call sites in api/archive.py, api/ingest.py, and api/insights.py
working via a local `active_archive_root as _active_archive_root`
alias rather than reaching into another module's private symbol.

Ref polylogue-exb
